feat: Merge NuvlaEdge with CLI
