+++ cli/common/urls.go
@@ -1,0 +1,18 @@
+package common
+
+// URL constants
+const (
+	// NuvlaEdgeBinaryURL NuvlaEdge binary url
+	NuvlaEdgeBinaryURL = "https://github.com/nuvlaedge/nuvlaedge-go/releases/download/%s/nuvlaedge-%s-%s-%s"
+
+	// NuvlaEdgeLatestConfTemplateURL NuvlaEdge latest configuration template url
+	NuvlaEdgeLatestConfTemplateURL = "https://raw.githubusercontent.com/nuvlaedge/nuvlaedge-go/main/config/template.toml"
+	NuvlaEdgeConfigTemplateURL     = "https://github.com/nuvlaedge/nuvlaedge-go/releases/download/%s/template.toml"
+
+	// NuvlaEdgeServiceURL NuvlaEdge service url
+	NuvlaEdgeServiceURL           = "https://github.com/nuvlaedge/nuvlaedge-go/releases/download/%s/nuvlaedge.service"
+	NuvlaEdgeLatestServiceFileURL = "https://raw.githubusercontent.com/nuvlaedge/nuvlaedge-go/main/installer/nuvlaedge.service"
+
+	// ReleasesURL NuvlaEdge releases url
+	ReleasesURL = "https://api.github.com/repos/{owner}/{repo}/releases"
+)
