+++ cli/errors/installer.go
@@ -1,0 +1,1 @@
+package errors
