+++ cli/installers/docker.go
@@ -1,0 +1,5 @@
+package installers
+
+type DockerInstaller struct {
+	dockerVersion string
+}
