+++ cli/types/installer.go
@@ -1,0 +1,10 @@
+package types
+
+type Installer interface {
+	Install() error
+	Start() error
+	Stop() error
+	Remove() error
+	Status() string
+	String() string
+}
