--- installer/nuvlaedge-cli/errors/installer.go
@@ -1,1 +1,0 @@
-package errors
