--- installer/nuvlaedge-cli/installers/docker.go
@@ -1,5 +1,0 @@
-package installers
-
-type DockerInstaller struct {
-	dockerVersion string
-}
