--- installer/nuvlaedge-cli/types/installer.go
@@ -1,10 +1,0 @@
-package types
-
-type Installer interface {
-	Install() error
-	Start() error
-	Stop() error
-	Remove() error
-	Status() string
-	String() string
-}
